requirements.txt
firebird-base==1.7.2
firebird-driver==1.10.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
protobuf==5.26.0
PyPika==0.48.9
pytest==8.1.1
python-dateutil==2.9.0.post0
six==1.16.0
